requirements.txt
fastapi==0.115.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-auth==2.38.0
google-generativeai==0.8.4
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
